create GuestUserExerciseRepository
